+++ src/main/java/dev/petshopsoftware/utilities/Util/InputChecker/LongChecker.java
@@ -14,19 +14,19 @@ public class LongChecker extends InputChecker<Long> {
 		}
 	}
 
-	public LongChecker min(int length, String message) {
+	public LongChecker min(long length, String message) {
 		return (LongChecker) check(input -> input >= length, message);
 	}
 
-	public LongChecker max(int length, String message) {
+	public LongChecker max(long length, String message) {
 		return (LongChecker) check(input -> input <= length, message);
 	}
 
-	public LongChecker range(int min, int max, String message) {
+	public LongChecker range(long min, long max, String message) {
 		return (LongChecker) check(input -> input >= min && input <= max, message);
 	}
 
-	public LongChecker mod(int mod, String message) {
+	public LongChecker mod(long mod, String message) {
 		return (LongChecker) check(input -> input % mod == 0, message);
 	}
 }
